+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## v12.0.0
+
+- **_Breaking Change_**: remove `TransformExtension.SetRotationX/Y/Z/W` and `SetLocalRotationX/Y/Z/W`.
+
 ## v11.3.0
 
 - Feat: `Array/IList.Fill` support `factory` as parameter to create reference types.